Cleanups
git-svn-id: svn+ssh://solsvn.oma.be/home/sidc/ssr/SWHV_repos/trunk/client/jhv-3d-wcs@7538 b4e469a2-07ce-4b26-9273-4d7d95a670c7

+++ src/jhv/src/org/helioviewer/jhv/gui/components/calendar/DayViewController.java
@@ -76,7 +76,7 @@ public class DayViewController implements CalendarViewController {
      */
     public void setDateOfCellValue(Object value) {
         if (value instanceof Integer) {
-            calendar.set(Calendar.DAY_OF_MONTH, ((Integer) value).intValue());
+            calendar.set(Calendar.DAY_OF_MONTH, (Integer) value);
         }
     }
 
@@ -105,7 +105,7 @@ public class DayViewController implements CalendarViewController {
 
         for (int i = 1; i <= numberOfDaysInMonth; i++) {
             cal.set(Calendar.DAY_OF_MONTH, i);
-            data[cal.get(Calendar.WEEK_OF_MONTH) - offset][cal.get(Calendar.DAY_OF_WEEK) - 1] = Integer.valueOf(i);
+            data[cal.get(Calendar.WEEK_OF_MONTH) - offset][cal.get(Calendar.DAY_OF_WEEK) - 1] = i;
         }
 
         return data;
